Move update user profiles to a separate saga
